% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWN_plotting_functions.R
\name{add_PT_buffer_parishes_borders}
\alias{add_PT_buffer_parishes_borders}
\title{Add Outline of the PT PWN Buffer Zone to an Existing Plot}
\usage{
add_PT_buffer_parishes_borders(backgrproj,
  buffer.zone.shp = "H:\\\\FISE\\\\forest\\\\CanopyHealthMonitoring\\\\PWN\\\\maps\\\\PT_buffer_zone\\\\FreguesiasZT_SHGM_LAEA.shp")
}
\arguments{
\item{backgrproj}{The projection of the current plot}

\item{buffer.zone.shp}{Shapefile of the buffer zone. Defaults to
H:\\FISE\\forest\\CanopyHealthMonitoring\\PWN\\maps\\PT_buffer_zone\\FreguesiasZT_SHGM_LAEA.shp}
}
\description{
Add Outline of the PT PWN Buffer Zone to an existing plot.

\code{add.PT.buffer.parishes.borders} plots European country borders on top of an existing plot, reading them from a .shp
}
