% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.r
\name{ADS100_NDVI}
\alias{ADS100_NDVI}
\title{Normalized Difference Vegetation Index simulating ADS100 sensor (ADS100_NDVI)}
\usage{
ADS100_NDVI(df, outp_fname = NULL, ...)
}
\arguments{
\item{df}{A data frame (or raster::brick) where columns (layers) represent reflectance measurements in a single wavelength, and
columns (layers) are named following Quantalab's conventions.}

\item{outp_fname}{In case the input is raster data, this is the optional output
filename to write the result to}

\item{...}{Arguments to be passed to get_band_of_wavelength, particularly band_txt, splitter, and i.}
}
\value{
A vector or rasterlayer with the value of the index
}
\description{
Calculate ADS100_NDVI
}
\note{
TO DO: the averaging across the broad band could be weighted using spectral response curve
}
\references{
http://w3.leica-geosystems.com/downloads123/zz/airborne/ADS100/brochures-datasheet/Leica_ADS100_DS_en.pdf
}
