% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_crowncells_quantile.r
\name{assign_crowncells_quantile}
\alias{assign_crowncells_quantile}
\title{Assign Polygon Cells Their Quantile Value In A Raster}
\usage{
assign_crowncells_quantile(r, crown_shp, outpname_r, plott = F,
  quantiles_to_keep = NULL, outpname_shp)
}
\arguments{
\item{r}{The single-layer raster object to draw values from, and that will serve as the spatial template for the output}

\item{crown_shp}{A polygon shapefile that overlays r}

\item{outpname_r}{A filename for the raster output}

\item{plott}{Do you want to plot the result for each crown? Default is F}

\item{quantiles_to_keep}{Integer vector with quantiles to keep in trimmed crowns output. Only considered if outpname_shp is not NULL.}

\item{outpname_shp}{The output filename for trimmed polygons. These retain only the part of the polygon with quantile values inside quantiles_to_keep.
If set to NULL, quantiles_to_keep is ignored and no vectorized output is written away. Default is NULL}
}
\value{
A raster file, with the same spatial attributes as r and quantile values between 0 and 100 for cells in polygons, and NA outside polygons
}
\description{
Overlay a polygon shapefile on a raster, so that each cell gets assigned its quantile value calculated from
all the raster values falling in that that particular polygon. Optionally, write away a polygon shapefile of the crowns where
only there portions with specified quantile values are kept.
}
\details{
Writes away  both 1) a raster file, with the same spatial attributes as r and quantile values between 0 and 100 for cells in polygons, and NA outside polygons, a
and 2) a shapefile with trimmed polygons
}
\note{
Write a line that checks wether a polygon intersects with the raster
}

