% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_growing_functions.r
\name{calc_neighbours_pix_nrs}
\alias{calc_neighbours_pix_nrs}
\title{Find Neigbour Cells in a Matrix}
\usage{
calc_neighbours_pix_nrs(nrow_, ncol_, include.diags = T)
}
\arguments{
\item{nrow_}{number of rows in the matrix}

\item{ncol_}{number of columns in the matrix}

\item{include.diags}{should diagonally neighbouring cells be considered? Default is TRUE.}
}
\value{
a matrix providing the linear cell number in the first row, and in following rows, the linear cell numbers of neighbour cells
}
\description{
When given a the nrow and ncol of a matrix, this function returns the cellnumbers of the cardinal neighbours of each cell.You can choose whether to also include diagonal neighbours using the include.diags flag.
}
\note{
called in region_growing_wrapper.r
}
