% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteness_functions.R
\name{calculate_remoteness_from_binary_raster}
\alias{calculate_remoteness_from_binary_raster}
\title{Calculate Remoteness Binary raster}
\usage{
calculate_remoteness_from_binary_raster(binary.raster, outer.diam.in.m = 10000
  * 2, inner.diam.in.m = 1500 * 2, plott = F)
}
\arguments{
\item{binary.raster}{A binary raster with dimensions in m}

\item{outer.diam.in.m}{The diameter of the disk inside which to calculated distance. In meters.}

\item{inner.diam.in.m}{The  diameter of the disk outside which to calculated distance. In meters.}

\item{plott}{Should the output be plotted?}
}
\value{
A raster with distance values.
}
\description{
When given a binary raster depicting a shape \code{calculate.remoteness.from.binary.raster} tells for each pixel, how far it is from the shape
}
