% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_growing_functions.r
\name{check_im_res}
\alias{check_im_res}
\title{Before Crown Delineation Check Image Resolution Consistency}
\usage{
check_im_res(r, max_radius)
}
\arguments{
\item{r}{a raster object}

\item{max_radius}{the maximum radius of a region to be grown}
}
\value{
the maximum radius expressed in number of pixels
}
\description{
For an image to be used in a region growing algorithm, check the image resolution for consistency with maximum region (ie crown) radius.
}
\note{
called in region_growing_wrapper.r
}
