% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWN_functions.R
\name{clip_polygons2rasters}
\alias{clip_polygons2rasters}
\title{Clip Spatialpolygons To The Extent Of A Raster}
\usage{
clip_polygons2rasters(SpatPol, r.fnames)
}
\arguments{
\item{SpatPol}{A SpatialPolygons(DataFrame)}

\item{r.fnames}{Names of raster files}
}
\value{
a SpatialPolygons(DataFrame) cut to the joint extent
}
\description{
given a SpatialPolygons(DataFrame) and a list of raster filenames, clips the Polygons to retain only what fall in
ONE OR MORE OF the rasters, i.e. to the joint extent of a set of rasters
}
