% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_growing_functions.r
\name{close_crown}
\alias{close_crown}
\title{Morphological Opening Operation}
\usage{
close_crown(outp., kern.size = 2)
}
\arguments{
\item{outp.}{a single-layer raster object with binary values}

\item{kern.size}{the size of the disc shaped kernel to use in the opening operation}
}
\value{
a SpatialPolygon object
}
\description{
A function to perform an opening operation on the crown
}
\note{
called in region_growing_wrapper.r
}
