% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_overview_of_rasters.r
\name{create_overview_of_rasters}
\alias{create_overview_of_rasters}
\title{Create A Polygon .shp Of Raster File Extents}
\usage{
create_overview_of_rasters(dirname)
}
\arguments{
\item{dirname}{Name of the directory to mine through}
}
\value{
A .shp file written to dirname and named images_overview.shp
}
\description{
Mine through a directory and make a single polygon .shp
where the polygon shows the extent of an image and there's an attribute with the filename. So far only .tif files are considered !
}
\examples{
\dontrun{
create_overview_of_rasters('H:/FISE/forest/CanopyHealthMonitoring/PWN/imagery/PT_Orto/Ortos_DistrCasteloBranco/')
}
}

