% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectral_indices.r
\name{CTRI2}
\alias{CTRI2}
\title{Carter Index 2}
\usage{
CTRI2(df, outp_fname = NULL)
}
\arguments{
\item{df}{A data frame where columns represent measurements in a single wavelength,
and columns are named following Quantalab's conventions}

\item{outp_fname}{In case the input is raster data, this is the optional output filename to write the result to}
}
\value{
A vector with the value of the index
}
\description{
Calculate Carter Index 2
}
\references{
x
}

