% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.r
\name{DMC_G2R}
\alias{DMC_G2R}
\title{Green_over_Red Index simulating DMC sensor (DMC_G2R)}
\usage{
DMC_G2R(df, outp_fname = NULL, ...)
}
\arguments{
\item{df}{A data frame (or raster::brick) where columns (layers) represent reflectance measurements in a single wavelength, and
columns (layers) are named following Quantalab's conventions.}

\item{outp_fname}{In case the input is raster data, this is the optional output
filename to write the result to}

\item{...}{Arguments to be passed to get_band_of_wavelength, particularly band_txt, splitter, and i.}
}
\value{
A vector or rasterlayer with the value of the index
}
\description{
Calculate DMC_G2R
}
\note{
TO DO: the averaging across the broad band could be weighted using spectral response curve
}
\references{
https://calval.cr.usgs.gov/JACIE_files/JACIE10/Presentations/ThurAM/Pagnutti_Mary_JACIE_2010_DigitalCameraCal.pdf
}
