% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_growing_functions.r
\name{eliminate_split_crown}
\alias{eliminate_split_crown}
\title{Clean up Split Polygons}
\usage{
eliminate_split_crown(crown_pol, testseed)
}
\arguments{
\item{crown_pol}{SpatialPolygons object that resulted from region growing. crown_pol might consist of multiple polygons}

\item{testseed}{SpatialPoints object from which crown_pol was grown}
}
\value{
a SpatialPolygons object containing only the polygon in which testseed falls
}
\description{
After the opening operation, the crown can get split into two. This function returns only the portion of the crown that the seed falls in
}
\note{
called in region_growing_wrapper.r
}
\references{
\url{http://w3facility.org/question/r-how-to-split-spatial-multipolygon-to-spatialpolygons-element/}
}
