% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_growing_functions.r
\name{ensure_uneven_r_dim}
\alias{ensure_uneven_r_dim}
\title{Before Crown Delineation Check Image Dimension Consistency}
\usage{
ensure_uneven_r_dim(r)
}
\arguments{
\item{r}{a raster object}
}
\value{
the input raster, cropped to uneven nrow and ncol if necessary
}
\description{
Ensure that a raster/brick has unven nr of pixels in x and y dimensions make sure the dimensions of the window are uneven as this is required for the region growing.
}
\note{
called in region_growing_wrapper.r
}
