% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_crown_growing_efficiency.r
\name{evaluate_crown_growing_efficiency}
\alias{evaluate_crown_growing_efficiency}
\title{Evaluate Efficiency Of Crown Delineation By CanHeMonR::grow_crowns}
\usage{
evaluate_crown_growing_efficiency(outp_crown_shp_filename)
}
\arguments{
\item{outp_crown_shp_filename}{File to which the results of CanHeMonR::grow_crowns were written}
}
\value{
graphs
}
\description{
Evaluate efficiency of crown delineation using CanHeMonR::grow_crowns. In particular, check whether the prob_cut parameter could be changed to save time.
}
