% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_classification_functions.R
\name{extract_mn_sd}
\alias{extract_mn_sd}
\title{Mean And SD From Polygon-on-Brick Overlay}
\usage{
extract_mn_sd(brick., polygon_Spat)
}
\arguments{
\item{brick.}{a raster brick}

\item{polygon_Spat}{a SpatialPolygons object}
}
\value{
A data frame with entries of bandwise mean ('Mean...') and sd ('Std...') for each polygon
}
\description{
Extracts means and sd for each band of a brick per each polygon in polygon_Spat.
This method achieves it in one extract pass over the polygons (functions in extract should return a single value)
}
