% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_growing_functions.r
\name{flag_valid_pix}
\alias{flag_valid_pix}
\title{Register Valid Pixels in a Matrix}
\usage{
flag_valid_pix(pixnr., prox_ind_)
}
\arguments{
\item{pixnr.}{The linear pixel number that should be registered}

\item{prox_ind_}{Matrix as produced by calc.neighbours.pix.nrs that registers pixel numbers in a matrix, their distance to the center of the matrix,
and the numbers of their neigbours}
}
\value{
the prox_ind_ matrix updated
}
\description{
A function that makes sure valid crown pixels are registered, so that their neigbours can be considered for inclusion in the crown as well
}
\note{
called in region_growing_wrapper.r
}
\seealso{
\code{calc.neighbours.pix.nrs}
}
