% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.r
\name{get_index_or_wavelength_from_brick}
\alias{get_index_or_wavelength_from_brick}
\title{Get Spectral Index or Chosen Wavelength From Brick}
\usage{
get_index_or_wavelength_from_brick(br, index_name_or_wavelength)
}
\arguments{
\item{br}{A multispectral raster brick. Layers should be named using the csic convention}

\item{index_name_or_wavelength}{Character or numeric. The name of a spectral index or the spectral wavelength,
in nanometers, to be extracted.}
}
\value{
A raster layer
}
\description{
Extract from a rasterband a particular wavelength, or spectral index
}
