% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TerraBella_scene_date.R
\name{get_TerraBella_scene_date}
\alias{get_TerraBella_scene_date}
\title{Get TerraBella Scene Date}
\usage{
get_TerraBella_scene_date(tifname)
}
\arguments{
\item{tifname}{Character vector of TerraBella filename with or without path.
e.g. 's02_20150525T110213Z_MS-0000020736-0000000000.tif'}
}
\value{
A dateframe with a Date object (rawdate) and a character
variable (sensornr) giving the sensor nr
}
\description{
Parse the date out of a TerraBella .tif filename
}
\examples{
\dontrun{
fnames <- list.files('X:/Imagery/Granadilla/Skybox')
fnames <- fnames[grep('tif',list.files('X:/Imagery/Granadilla/Skybox'))]
fnames <- fnames[substr(fnames,nchar(fnames)-3,nchar(fnames)) == ".tif"]
imdates <- get_TerraBella_scene_date(fnames)
win.graph(w = 12, h=5)
graphics::hist(imdates, breaks = 'months', format = "\%b \%Y", freq = T)
axis.Date(4, at = "2015-01-01")
}
}
