% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_growing_functions.r
\name{grow_crown}
\alias{grow_crown}
\title{Grow a Spectrally Homogenous Object}
\usage{
grow_crown(testim, prox_ind, startclus, prob_cut)
}
\arguments{
\item{testim}{A single or multi-band raster image to grow a region in}

\item{prox_ind}{A matrix, as produced by Proximity.indices, that details proximity to the center and neighbours for all the cells in testim}

\item{startclus}{The number of cells at the center of the testim that should be taken for granted as belonging to the region}

\item{prob_cut}{The  cutoff to apply when considering the probability that a new pixel at the edge of the region, belongs to the region. The value should fall between 0 an 100. Higher values tend to produce larger regionss}
}
\value{
A list containing $crown: binary raster of the region grown, and matching the attributes of testim and $startcluster: the startcluster size required to produce a non-singular covariance matrix (required to calculate Mahalanobis distance to the startcluster)
}
\description{
Grow a spectrally homogenous object from the center of a raster object outwards, based on Mahalonobis distance criteria
}
\note{
called in region_growing_wrapper.r
}
\references{
\url{http://w3facility.org/question/r-how-to-split-spatial-multipolygon-to-spatialpolygons-element/}
}
