% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_sampling.R
\name{histnd}
\alias{histnd}
\title{Multidimensional Stratification}
\usage{
histnd(df, quants = c(0, 0.05, 0.2, 0.8, 0.95, 1))
}
\arguments{
\item{df}{A dataframe with continuous variables}

\item{quants}{Quantiles to use as breakpoints in stratification}
}
\value{
A vector of length nrow(df) with a unique value for each stratum
}
\description{
Calculate strata along multiple variables
}
\note{
for the time being strata are cut for each variable at the quantiles.
Default quantiles are c(0,0.075,0.30,0.70,0.925,1)
#http://stackoverflow.com/questions/32614736/multidimensional-histogram-with-r
}
