% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWN_functions.R
\name{identify_n_neighbours}
\alias{identify_n_neighbours}
\title{Find A Polygon's Neighbour Polygons}
\usage{
identify_n_neighbours(Spat.polygons, n, append = F,
  col.prefix = "IDneighb_")
}
\arguments{
\item{Spat.polygons}{A SpatialPolygons(DataFrame)}

\item{n}{The nr of neighbouring polygons to detect}

\item{append}{Should the output be appended to the data in the SpatialPolygonsDatafrmae}

\item{col.prefix}{A prefix to use in the output of the columns that include hold the IDs of the neighbours}
}
\value{
a SpatialPolygonsDataframe with a column with unique IDs and n columns holding the IDs of the n closest neighbours
}
\description{
Identify for each element in a SpatialPolygons object, its n closest neighbours.
}
\examples{
\dontrun{
Decline.cluster <- shapefile('E:/beckpie/PWN/vector_data/Decline_N_tree_clusters.shp')
tt <- identify_n_neighbours(Spat.polygons = Decline.cluster,n=2,append=T)
}
}
