% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWN_functions.R
\name{indiv_bands}
\alias{indiv_bands}
\title{Get A List Into A df For Easier Manipulation}
\usage{
indiv_bands(x, dfnames)
}
\arguments{
\item{x}{a list}

\item{dfnames}{a vector with the names of datasets in the lists}
}
\value{
a dataframe
}
\description{
a function to get a list into a df for easier manipulation and ggplot display
}
