% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.r
\name{LIC1}
\alias{LIC1}
\title{Lichtenthaler Index 1}
\usage{
LIC1(df, outp_fname = NULL, ...)
}
\arguments{
\item{df}{A data frame where columns represent measurements in a single wavelength,
and columns are named following Quantalab's conventions}

\item{outp_fname}{In case the input is raster data, this is the optional output filename to write the result to}
}
\value{
A vector with the value of the index
}
\description{
Calculate Lichtenthaler Index 1
}
\references{
Lichtenthaler et al 1996
}

