% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictor_txt_files.r
\name{make_predictor_txt_files}
\alias{make_predictor_txt_files}
\title{Create a tx file with predictors}
\usage{
make_predictor_txt_files(r_train_dir, tile, data_outp_dir, parallel, nWorkers)
}
\arguments{
\item{r_train_dir}{A directory where .tifs for the prediction can be found}

\item{tile}{Character vector. Names of tile(s) to run. 'ALL will run all tiles in r_train_dir. Default is 'ALL'}

\item{data_outp_dir}{The folder and filename prefix to save the sampled data to. No data is saved is data_outp_dir is NULL. Default is NULL.}

\item{parallel}{Should the code be run in parallel using the doParallel package? Default is FALSE.}

\item{nWorkers}{If running the ocde in parallel, how many workers should be used? Default is 4.}
}
\value{
A txt file with the paths of the tiles to predict in the function run_sicktree_model_multi_tile.r
}
\description{
Create a txt file with the paths of the predictors to be used in the run_sicktree_model
}
\examples{
\dontrun{
}
}
