% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWN_plotting_functions.R
\name{mask_to_PT_buffer_zone}
\alias{mask_to_PT_buffer_zone}
\title{Mask a Raster File to Only Retain Data within the PT PWN Buffer Zone}
\usage{
mask_to_PT_buffer_zone(r,
  buffer.zone.shp = "H:\\\\FISE\\\\forest\\\\CanopyHealthMonitoring\\\\PWN\\\\maps\\\\PT_buffer_zone\\\\FreguesiasZT_SHGM_LAEA.shp",
  filename = NA)
}
\arguments{
\item{r}{A raster object}

\item{buffer.zone.shp}{Shapefile of the buffer zone. Defaults to
H:\\FISE\\forest\\CanopyHealthMonitoring\\PWN\\maps\\PT_buffer_zone\\FreguesiasZT_SHGM_LAEA.shp}

\item{filename}{Optional filename to write the output file to}
}
\description{
\code{mask.to.PT.buffer.zone} masks a raster file to only retain data within the PT PWN buffer zone.
}
