% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWN_functions.R
\name{merge_pols_intersected_by_lines}
\alias{merge_pols_intersected_by_lines}
\title{Merge Polygons Intersected By Lines}
\usage{
merge_pols_intersected_by_lines(spat_pols, spat_lines, original_res = 0.15)
}
\arguments{
\item{spat_pols}{A SpatialPolygons object}

\item{spat_lines}{A SpatialLines object, e.g. as created by tile_raster_extent}

\item{original_res}{The original resolution of the input image the Polygons were calculated from. Default = 0.15}
}
\value{
A SpatialPolygons object
}
\description{
When an object-identification function is run on raster image in tiled mode (for memory management or paralel processing),
The resulting objects are intersected by lines that correspond to the tile borders. This function allows you to mergein a SpatialPolygons
object all the polygons that are split by a set of SpatialLines
}
\seealso{
tile_raster_extent
}
