% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.r
\name{NDVI}
\alias{NDVI}
\title{Normalized Difference Vegetation Index}
\usage{
NDVI(df, outp_fname = NULL)
}
\arguments{
\item{df}{A data frame where columns represent measurements in a single wavelength, and columns are named following Quantalab's conventions}

\item{outp_fname}{In case the input is raster data, this is the optional output filename to write the result to}
}
\value{
A vector with the value of the index
}
\description{
Calculate NDVI
}
\examples{
\dontrun{
#Calculate the NDVI from a 6 band multispectral image
my_ms_data <- raster::brick('H:/FISE/forest/CanopyHealthMonitoring/PWN/flights_final/150727_mca/150727_mca.bsq')
my_NDVI <- NDVI(my_ms_data)
}
}
\references{
Rouse et al. 1974
}

