% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_sampling.R
\name{pairsplot}
\alias{pairsplot}
\title{A Customized Pairs Plot}
\usage{
pairsplot(mydat, cols2ignore = c("primalX", "primalY", "npx_t_b", "nstrtcl",
  "probcut"), cols2include = NA)
}
\arguments{
\item{mydat}{A dataframe}

\item{cols2ignore}{Character vector. Columns in the attribute table that should be ignored in the stratification.
Default is c('primalX','primalY','npx_t_b','nstrtcl','probcut') which are legacy attributes.
Furthermore columns of which the name starts with 'x' are ingored automatically.}

\item{cols2include}{Character vector. Columns in the attribute table that should be included in the stratification.
Default is NA, which means all columns, except for those in cols2ignore are included. cols2ignore takes precedence over cols2include}
}
\value{
a pairs plot
}
\description{
Summarize a data set visually using a pairs plot
}
