% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons_to_seeds.r
\name{polygons_to_seeds}
\alias{polygons_to_seeds}
\title{Convert Polygons To Seeds}
\usage{
polygons_to_seeds(rough_crowns_shp_fname, seed_placement = "centroid",
  image_fname, index_name = "NDVI", parallel = F, nWorkers = 2,
  outp_shp_fname, append = F, bandnames = NULL)
}
\arguments{
\item{rough_crowns_shp_fname}{Character. Filename for the of the input polygon shapefile roughly depicting crowns.}

\item{seed_placement}{How should polygons be converted to points. Options are 'centroid', which
uses the centroid of the polygon, and 'maxval', which uses the location in the polygon where index_name reaches its maximum. Defaults is 'centroid'}

\item{image_fname}{Optional. Filename of the image to run extract spectral bands or indices from to provide context.
Only considered if seed_placement is 'maxval'.}

\item{index_name}{Character (for a spectral index) or numeric (for an individual wavelength).
The spectral index or wavelength on which to perform the contextual evaluation. If a wavelength is provided, it should be expressed in nm.
Only considered if seed_placement is 'maxval'.
Default is "NDVI".}

\item{parallel}{Logical. Would you like the tiles to be processed in parallel?  Default is False.}

\item{nWorkers}{If running the code in parallel, how many workers should be used? Default is 2.}

\item{outp_shp_fname}{Character. Filename for the output point shapefile.}

\item{append}{Logical. Should the data in outp_shp_fname be appendend with the new results? Default is False.}

\item{bandnames}{Character. In case the bands aren't named according to wavelength and following csic convention, they
can be provided. Default is NULL in which cases bandnames are read from the image file and csic naming convention is assumed.}
}
\value{
A SpatialPoints object
}
\description{
Convert SpatialPolygons to points, in a context-specific manner
}
\seealso{
watershed_tree_delineation
}
