% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_growing_functions.r
\name{proximity_indices}
\alias{proximity_indices}
\title{Make a distance-to-center matrix}
\usage{
proximity_indices(nrow_ = 3, ncol_ = 6)
}
\arguments{
\item{nrow_}{nr of rows of the matrix}

\item{ncol_}{nr of columsn of the matrix}
}
\value{
a 2 row matrix where row 1: cell nr, and row 2: distance to center measured in cellnrs
}
\description{
For a matrix, order cells by distance to the center.
}
\note{
called in region_growing_wrapper.r
}
