% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWN_functions.R
\name{remove_all_NA}
\alias{remove_all_NA}
\title{Check An Object For NAs}
\usage{
remove_all_NA(df)
}
\arguments{
\item{df}{An object}
}
\value{
The object if has any non-na elements, else NULL
}
\description{
Check if an object contains only NAs
}
