% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_small_pols.r
\name{remove_small_pols}
\alias{remove_small_pols}
\title{Remove Small Polygons}
\usage{
remove_small_pols(spatpols, minsize, outname = NULL)
}
\arguments{
\item{spatpols}{A SpatialPolygons object}

\item{minsize}{numeric Any polygons below this size will be removed}

\item{outname}{character. Optional filename to write the result to}
}
\value{
A SpatialPolygons(DataFrame) object
}
\description{
Remove small polygons from a SpatialPolygons object
}
