% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_classification_functions.R
\name{rF_predictions_to_polygons}
\alias{rF_predictions_to_polygons}
\title{Run An Existing rF In Predictive Mode}
\usage{
rF_predictions_to_polygons(rF_model, segments_shp, WV2image, append_shp = F,
  target_class = "EG", outp_shp)
}
\arguments{
\item{rF_model}{an .rdata file that holds an rF classifier named "rF"}

\item{segments_shp}{a polygon shapefile of landscape elements that need to be classified}

\item{WV2image}{an image to extract mn and sd from, for each polygon, as input to the rF}

\item{append_shp}{should the original fields of segments.shp be retained in the output?}

\item{target_class}{(for rF classifiers only) if a target.class is given,
only polygons predicted by the rF to be this class will be retained in the output}

\item{outp_shp}{name for the output shapefile}
}
\value{
a polygon shapefile, akin to segments.shp that has one field, holding with the output of the rF
}
\description{
Run an existing rF in predictive mode on mn and sd for each polygon, extracted from a raster
}
