% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sicktree_model_multi_tile.r,
%   R/run_sicktree_model_multi_tile_BDEOSS.r
\name{run_sicktree_model_multitile}
\alias{run_sicktree_model_multitile}
\alias{run_sicktree_model_multitile}
\title{Run a saved MaxEnt model in predictive mode on a tile of image data}
\usage{
run_sicktree_model_multitile(predictors_dir, txt_dir, fname_predictors_txt,
  MaxEntmodel_dir, fname_MaxEntmodel_r, output_dir)

run_sicktree_model_multitile(predictors_dir, txt_dir, fname_predictors_txt,
  MaxEntmodel_dir, fname_MaxEntmodel_r, output_dir)
}
\arguments{
\item{predictors_dir}{Path to the tile that we want to predict}

\item{txt_dir}{Path to txt file with the predictors}

\item{fname_predictors_txt}{Textfile specifying the predictors (ie covariates) for the model as image filenames in the correct order}

\item{MaxEntmodel_dir}{Directory where the MaxEnt model file is held}

\item{fname_MaxEntmodel_r}{Filename of the MaxEnt model saved in rds format (see ?readRDS)}

\item{output_dir}{Output directory for the tif}

\item{predictors_dir}{Direcotry where predictor layers are held}

\item{txt_dir}{Path where a txt file listing predictor layers is held}

\item{fname_predictors_txt}{Textfile specifying the predictors (ie covariates) for the model as image filenames in the correct order}

\item{fname_MaxEntmodel_r}{Filename of the MaxEnt model saved in rds format (see ?readRDS)}

\item{output_dir}{Directory to write the output to}
}
\value{
Saves class-specific distribution models as raster images, using image layers as inputs

Saves class-specific distribution models as raster images, using image layers as inputs
}
\description{
Run a saved MaxEnt model in predictive mode on a tile of image data

Run a saved MaxEnt model in predictive mode on a tile of image data
}
\note{
Run in 32-bit R installation.
}
\examples{
\dontrun{
run_sicktree_model_multitile(fname_predictors_txt = file.path(model_dir <- "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/classification_temp/",'predictors_pt606000_4401000.txt'),
                            fname_MaxEntmodel_r = file.path(model_dir <- "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/classification_temp/", 'Pb.rdsdata'),
                            fname_output_tif =  file.path(model_dir <- "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/classification_temp/",'MaxEnt_Pb_pt606000_4401000.tif'))
run the tile for which you had a good model trained on that tile only.
the difference was that that earlier model sampled from circles around the points.
model_dir <- "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/classification_temp/"
run_sicktree_model_multitile(fname_predictors_txt = file.path(model_dir <- "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/classification_temp/",'predictors_pt617000_4404000.txt'),
                            fname_MaxEntmodel_r = file.path(model_dir <- "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/classification_temp/", 'samp10_Pb.rdsdata'),
                            fname_output_tif =  file.path(model_dir <- "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/classification_temp/",'MaxEnt_Pb_pt617000_4404000_100.tif'))
}

\dontrun{
run_sicktree_model_multitile(
 txt_dir = "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/classification_temp/",
 fname_predictors_txt = "predictors_pt617000_4404000.txt",
 predictors_dir = "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/RGBN_LUT",
 MaxEntmodel_dir = "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/classification_temp",
 fname_MaxEntmodel_r = "samp10_Pb.rdsdata",
 output_dir = "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/Imagery/Portugal/DMC/ortophotos_22122016/classification_temp",
)
}
}
