% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_false_positives.r
\name{sample_false_positives}
\alias{sample_false_positives}
\title{Sample false positives from a set of binary images}
\usage{

  sample_false_positives(r_pred_dir = "//ies.jrc.it/h03/CANHEMON/H03_CANHEMON/TMP_FOR_PIETER/test_docker3_out_processed/",
  tile = "ALL", points_to_avoid = NA, mindistance_to_presence = 80,
  perc = 0.01, max_n = 100, parallel = F, nWorkers = 4)
}
\arguments{
\item{r_pred_dir}{A directory where binary .tifs predicting presence as 1 and absence as 0 can be found for multiple tiles}

\item{tile}{Character vector. Names of tile(s) to run. 'ALL will run all tiles in r_pred_dir. Default is 'ALL'}

\item{points_to_avoid}{A SpatialPoints object describing areas that should be excluded from sampling. These are typically
true presence points Optional.}

\item{mindistance_to_presence}{The distance, in map units, that the sampled false positives should be from the true positives provided
in points_to_avoid.}

\item{perc}{The percentage of the available false positive pixels in each image to sample. Default is 1.}

\item{max_n}{The absolute maximum of locations to extract per image. This value overrides the proportion.}

\item{parallel}{Should the code be run in parallel using the doParallel package? Default is FALSE.}

\item{nWorkers}{If running the ocde in parallel, how many workers should be used? Default is 4.}
}
\description{
Sample false positives from a set of binary images
}
