% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_sampling.R
\name{samplestratified}
\alias{samplestratified}
\title{Take A Stratified Sample Of A Data Set}
\usage{
samplestratified(mydat, nsamples, strats)
}
\arguments{
\item{mydat}{A data frame with continous variables}

\item{nsamples}{The number of samples you'd like to take}

\item{strats}{A vector of length nrow(mydat) with a unique value for each stratum}
}
\value{
A dataframe that is a subset of mydat, with nrows close to nsamples
}
\description{
Take a stratified sample of chosen size from a data
}
