% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_sampling.R
\name{samplestratified_shp}
\alias{samplestratified_shp}
\title{Take A Stratified Sample Of A .SHP With Spectral Indices As Attributes}
\usage{
samplestratified_shp(shp_fname, shp_outname, nsamples, plott = T,
  cols2ignore = c("primalX", "primalY", "npx_t_b", "nstrtcl", "probcut"),
  cols2include = NA)
}
\arguments{
\item{shp_fname}{A shp point file with spectral indices}

\item{shp_outname}{The filenmae for the output shapefile of the sample}

\item{nsamples}{The number of samples you'd like to take}

\item{plott}{Logical. Should diagnostic plots be produced? Default is F.}

\item{cols2ignore}{Character vector. Columns in the attribute table that should be ignored in the stratification.
Default is c('primalX','primalY','npx_t_b','nstrtcl','probcut') which are legacy attributes.
Furthermore columns of which the name starts with 'x' are ingored automatically.}

\item{cols2include}{Character vector. Columns in the attribute table that should be included in the stratification.
Default is NA, which means all columns, except for those in cols2ignore are included. cols2ignore takes precedence over cols2include}
}
\value{
A shapefile that is a subset of the input, with nrows close to nsamples
}
\description{
Take a stratified sample of chosen size from a .shp data set of spectral indices
}
