% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices_for_crowns.r
\name{spectral_indices_for_crowns}
\alias{spectral_indices_for_crowns}
\title{Calculate Spectral Indices For Individual Tree Crowns}
\usage{
spectral_indices_for_crowns(crown_shp, r_fname, index_names = NULL,
  bandnames = NULL,
  outname = "E:\\\\beckpie\\\\temp\\\\outpout_example2.shp")
}
\arguments{
\item{crown_shp}{Shapefile depicting crowns}

\item{r_fname}{Raster file of the image to calculate the spectral index from.
The bands of the raster should be named following the Quantalab convention *X700.000000.Nanometers*}

\item{index_names}{Character vector with the names of a spectral indices implemented as a function in CanHeMonR. e.g. "G_over_B".
The index_name 'all_bands' can be included to also return individual band values}

\item{bandnames}{Character. In case the bands aren't named according to wavelength and following csic convention, they
can be provided. Default is NULL in which cases bandnames are read from the image file and csic naming convention is assumed.}

\item{outname}{Filename for a shapefile of the output}
}
\value{
A point shapefile with attribute table, and the points placed at the centroid of each crown.
The attribute table of the shp file maintains the attribute table of the crown but adds a column for each spectral index.
}
\description{
Calculate one or more spectral indices for individual
tree crowns depicted in a polygon shapefile and a provided raster image and write the output to a .shp
}

