% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWN_functions.R
\name{subtract_control_median}
\alias{subtract_control_median}
\title{Subtract Band And Polygon Specific Median From Digital Numbers}
\usage{
subtract_control_median(df)
}
\arguments{
\item{df}{A data frame as generated by indiv_bands, which inclusdes the variables
ID, dfname, band, DN.s}
}
\value{
A data frame with the variable DN.c added
}
\description{
A function to subtract from DN.s values (per pixel, data set and band),
the median in matching control polygons. The control polygons are usually the n. closest neighbours
}
