% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.r
\name{TCARI_over_OSAVI}
\alias{TCARI_over_OSAVI}
\title{Transformed Chlorophyll Absorption in Reflectance Index Over Optimised Soil-Adjusted Vegetation Index}
\usage{
TCARI_over_OSAVI(df, outp_fname = NULL, ...)
}
\arguments{
\item{df}{A data frame where columns represent measurements in a single wavelength,
and columns are named following Quantalab's conventions}

\item{outp_fname}{In case the input is raster data, this is the optional output filename to write the result to}

\item{...}{Arguments to be passed to get_band_of_wavelength, particularly band_txt, splitter, and i.}
}
\value{
A vector with the value of the index
}
\description{
Calculate Transformed Chlorophyll Absorption in Reflectance Index Over Optimised Soil-Adjusted Vegetation Index
}
\note{
TCARI is mostly pigment related but, at least in crops, is also sensitive to structure. Hence
TCARI/OSAVI estimates chlorohyll content better in eg mais crops over a range of LAI values
}
\references{
Haboudane et al. 2002
}
