% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWN_functions.R
\name{tile_raster_extent}
\alias{tile_raster_extent}
\title{Tile A Raster's Extent}
\usage{
tile_raster_extent(r, max_pixs)
}
\arguments{
\item{r}{A raster object}

\item{max_pixs}{Integer The maximum number of pixels that each tile can contain.}
}
\value{
A list of extent objects
}
\description{
Tile the extent object of a raster object for tiled processing
}
\seealso{
merge_pols_intersected_by_line
}
