% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_classification_functions.R
\name{train_rF_classifier_on_polygons}
\alias{train_rF_classifier_on_polygons}
\title{Calibrates A Supervised Classification Of Multiband Imagery Data}
\usage{
train_rF_classifier_on_polygons(polygon_SpatDF, calibration_field,
  training_brick, outp_rdata_file)
}
\arguments{
\item{polygon_SpatDF}{SpatialPolygonsDataFrame}

\item{calibration_field}{The field in polygon_SpatDF that contains the classes and serves as response variable for the model.}

\item{training_brick}{Raster brick from which values are extracted, as predictor variables}

\item{outp_rdata_file}{Output filename for an rdata object containing the rF model}
}
\value{
rF model, also written away to file
}
\description{
Calibrates a supervised classification of multiband imagery data, using reference data (presence/absence) in a polygon shape file
}
\note{
TO DO: Provide options with classification (rF, where you have pres/abs training data)
and similarity, where you only have pres training data that uses  maxent
DOES CROPPPING TO COMMON EXTENT SAVE TIME ?
}
