% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_TerraBella_indices.r
\name{calculate_TerraBella_indices}
\alias{calculate_TerraBella_indices}
\title{Calculate TerraBella spectral indices}
\usage{
calculate_TerraBella_indices(inp_tif_fnames, index_names, outp_dir,
  overwrite = T)
}
\arguments{
\item{inp_tif_fnames}{Filenames of an TerraBella tif files}

\item{index_names}{Character vector of spectral indices (see CanHeMonR::spectral_indices)
you want to have calculated. Remember TerraBella is 4 broad bands only}

\item{outp_dir}{Directory to write output to}

\item{overwrite}{Should existing output be overwritten? Default is T.}
}
\value{
a raster brick with the chosen indices and a file for each index,
written away to the output directory
}
\description{
When given a TerraBella .tif file, calculate chosen spectral indices
}
\note{
TO DO: adapt to work in parallel
}
\examples{
\dontrun{
#calculate NDVI for all the TerraBella scenes of Granadilla
raster::rasterOptions(progress='text')
raster::rasterOptions(tmpdir = 'E:/beckpie/temp/Raster_temp/')
inpdir <- 'X:/Imagery/Granadilla/Skybox/dark_object_corrected'
fnames <- list.files(inpdir)
fnames <- fnames[grep('.tif',fnames)]
fnames <- fnames[substr(fnames,nchar(fnames)-3,nchar(fnames))==".tif"]
fnames <- file.path(inpdir, fnames)

calculate_TerraBella_indices(inp_tif_fnames = fnames, index_names='NDVI',
                           outp_dir ="X:/Imagery/Granadilla/Skybox/dark_object_corrected/spectral_indices")
}
}
