% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_sicktree_model.r
\name{calibrate_sicktree_model}
\alias{calibrate_sicktree_model}
\title{Calibrate vegetation distribution models}
\usage{
calibrate_sicktree_model(vuln_classes = "ALL", training_df, model_outp_dir)
}
\arguments{
\item{vuln_classes}{A list of the classes you want to model.
The list can contain one or more vectors. Each vector represents a seperate vegetation class and response variable for the model
and the vector elements are synonyms used to describe that class.
The fist place in each vector will be used in the output name used to store the calibrated model, so it should not contain spaces.
The other places should appear as attributes in the field 'field_name' of Pols.}

\item{training_df}{Read the rds data that has been saved on the sampling step, see example}

\item{model_outp_dir}{The folder and filename prefix to save the model objects to}
}
\value{
Saves class-specific distribution models, using image layers as inputs
}
\description{
For each class in .shp polygon file, calibrate a distribution model using a raster brick as predictors
}
\note{
Run in 32-bit R installation. Do you need a 'require(rJava)?'. Implement optional parallel
}
\examples{
\dontrun{
}

#' Laura
}
