% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectral_indices.r
\name{get_band_of_wavelength}
\alias{get_band_of_wavelength}
\title{Get The Measurements Made In A Chosen Wavelength}
\usage{
get_band_of_wavelength(spec_df, wavelength_in_nm)
}
\arguments{
\item{spec_df}{Dataframe or RasterBrick of spectral measurements, with a column per band/spectral wavelength,
and columns (layers if spec_df is RasterBrick) named following Quantalab's convention.}

\item{wavelength_in_nm}{Integer The wavelength of interest, in nm.}
}
\value{
the column or rasterlayer with measurements in the chosen wavelength
}
\description{
Extract from a dataframe where columns are measurements in individual wavelengths,
and named following Quantalab's convention, the data that are closest to a chosen wavelegth.
}
\note{
TO DO: adapt to keep decimals in wavelength value written in layer name
}

