% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_indices.r
\name{NDVI}
\alias{NDVI}
\title{Normalized Difference Vegetation Index (NDVI)}
\usage{
NDVI(df, outp_fname = NULL, ...)
}
\arguments{
\item{df}{A data frame (or raster::brick) where columns (layers) represent reflectance measurements in a single wavelength, and
columns (layers) are named following Quantalab's conventions.}

\item{outp_fname}{In case the input is raster data, this is the optional output
filename to write the result to}

\item{...}{Arguments to be passed to get_band_of_wavelength, particularly band_txt, splitter, and i.}
}
\value{
A vector with the value of the index
}
\description{
Calculate NDVI
}
\examples{
\dontrun{
#Calculate the NDVI from a 6 band multispectral image
my_ms_data <- raster::brick('H:/FISE/forest/CanopyHealthMonitoring/PWN/flights_final/150727_mca/150727_mca.bsq')
my_NDVI <- CanHeMonR::NDVI(my_ms_data)
}
}
\references{
Rouse et al. 1974
}
