% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_based_seed_filter.R
\name{shadow_based_seed_filter}
\alias{shadow_based_seed_filter}
\title{Filter Seeds That Lack Shadows}
\usage{
shadow_based_seed_filter(seed_shp_fname, image_fname, shadow_mask, x_shift,
  y_shift, shadow_diameter = 6)
}
\arguments{
\item{seed_shp_fname}{Character. Filename for the input point shapefile.}

\item{image_fname}{Filename of the image to run the tree detection on}

\item{shadow_mask}{List. Uneven positions should provide index names or wavelength numbers. Even positions in the list
should provide the cut-off value of the spectral index, below which, pixels are considered shadow.
Default is list('NDVI',0,1).}

\item{x_shift}{Numeric. Horizontal shift from crown to shadow in x_direction.}

\item{y_shift}{Numeric. Horizontal shift from crown to shadow in y_direction.}

\item{shadow_diameter}{Numeric. Maximum diameter of a crown. Default is 6 m.}

\item{outp_seed_shp_fname}{Character. Filename for the output point shapefile.}
}
\value{
A SpatialPoints object
}
\description{
Remove spatial points that don't appear to have a tree-like shadow in a chosen spectral index or band.
}
\examples{
\dontrun{
shadow_based_seed_filter(
seed_shp_fname = 'C:/Users/pieterbeck/Documents/temp/test_treedetection_seeds.shp',
image_fname = image_fname = 'E:/FISE/forest/CanopyHealthMonitoring/PWN/flights_final/150727_mca/150727_mca.bsq',
shadow_mask = list(800, 801),
x_shift = -4,
y_shift = 4,
shadow_diameter = 5,
outp_seed_shp_fname = 'C:/Users/pieterbeck/Documents/temp/test_treedetection_seeds_shadow_masked.shp' )
}
}

