% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_test_fun.r
\name{test_fun}
\alias{test_fun}
\title{Test function for grid}
\usage{
test_fun(x)
}
\arguments{
\item{x}{blala}
}
\value{
A raster layer
}
\description{
Extract from a rasterband a particular wavelength, or spectral index
}
